pyproject.toml
[project]
name = "mqtt-data-bridge"
version = "0.1.0"
description = ""
authors = [
    {name = "Leandro Mario",email = "leandro.mario@lactec.com.br"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "pytest (>=9.0.1,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
